restructured mock fetcher and mapper tests
